Initial commit
Given a stream of N data points: (score, category), find the top k
points with the highest scores, and calculate the rank and statistical
significance of categories of the top k points. Stream is the file
input with score and category separated by a tab, a binary max heap
stores the top k data points and stores the information pertaining to
population size, success states in population, number of draws, and the
number of successes in the top k data points. Once those values have
been gathered, the hyper-geometric distribution is calculated for each
category and merge sort is used to sort the categories based on the
hyper-geometric distribution from highest to lowest.